# nextjs-frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/emails/gmail/connect-url": {
      "get": {
        "tags": [
          "email"
        ],
        "summary": "Get Gmail Connect Url",
        "operationId": "get_gmail_connect_url",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GmailConnectUrlResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/emails/gmail/status": {
      "get": {
        "tags": [
          "email"
        ],
        "summary": "Get Gmail Status",
        "operationId": "get_gmail_status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GmailConnectionStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/emails/gmail/oauth/callback": {
      "get": {
        "tags": [
          "email"
        ],
        "summary": "Gmail Oauth Callback",
        "operationId": "gmail_oauth_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/emails/gmail/triage-connected": {
      "post": {
        "tags": [
          "email"
        ],
        "summary": "Triage Connected Gmail",
        "operationId": "triage_connected_gmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectedGmailTriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GmailMcpTriageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/emails/gmail/webhook": {
      "post": {
        "tags": [
          "email"
        ],
        "summary": "Gmail Webhook",
        "operationId": "gmail_webhook",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GmailWebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GmailWebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectedGmailTriageRequest": {
        "properties": {
          "search_query": {
            "type": "string",
            "title": "Search Query",
            "description": "Gmail search query for the messages OpenAI should triage.",
            "default": "newer_than:1d"
          },
          "max_results": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Results",
            "default": 50
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          },
          "high_label": {
            "type": "string",
            "title": "High Label",
            "default": "Priority/High"
          },
          "medium_label": {
            "type": "string",
            "title": "Medium Label",
            "default": "Priority/Medium"
          },
          "low_label": {
            "type": "string",
            "title": "Low Label",
            "default": "Priority/Low"
          },
          "user_focus": {
            "type": "string",
            "title": "User Focus",
            "default": "Prioritize direct human requests, legal or finance issues, deadlines, customer issues, action items, and emails with specific dates or times. Promotional emails should usually be low priority."
          }
        },
        "type": "object",
        "title": "ConnectedGmailTriageRequest"
      },
...
      "GmailCategory": {
        "type": "string",
        "enum": [
          "primary",
          "promotions",
          "updates"
        ],
        "title": "GmailCategory"
      },
      "GmailConnectUrlResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "GmailConnectUrlResponse"
      },
      "GmailConnectionStatusResponse": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "email_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Address"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/GmailCategory"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "connected"
        ],
        "title": "GmailConnectionStatusResponse"
      },
...
      "GmailMcpTriageResponse": {
        "properties": {
          "result": {
            "type": "string",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "GmailMcpTriageResponse"
      },
...
      "GmailWebhookMessage": {
        "properties": {
          "data": {
            "type": "string",
            "title": "Data"
          },
          "messageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messageid"
          },
          "publishTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishtime"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "GmailWebhookMessage"
      },
      "GmailWebhookPayload": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/GmailWebhookMessage"
          },
          "subscription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "GmailWebhookPayload"
      },
      "GmailWebhookResponse": {
        "properties": {
          "processed": {
            "type": "boolean",
            "title": "Processed"
          },
          "email_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Address"
          },
          "pulled": {
            "type": "integer",
            "title": "Pulled",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "processed"
        ],
        "title": "GmailWebhookResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}